[audio_classification] Fix indentation and change filename
